Spike looking into a possible `enqueuePriority` task modifier.
This modifier would allow you to pass a `sortFunc` which is used to
sort the queue directly before taking any tasks off it. This means that
you can add a bunch of tasks to a queue but mark some as more important
than others.

The contents of this commit are just a spike based of some advice from
@machty in slack. If this general approach makes sense then I could
try to make sense out of the tests and round this out into a more
complete PR...